Splits up `shell_command.py`
